Extract agent graph record mapping
