+++ web/src/server/api/routers/traces.ts
@@ -59,10 +59,8 @@ import { TRPCError } from "@trpc/server";
 import { createBatchActionJob } from "@/src/features/table/server/createBatchActionJob";
 import { throwIfNoEntitlement } from "@/src/features/entitlements/server/hasEntitlement";
 import { sanitizeLegacyTracingSearch } from "@/src/features/traces/server/legacyIoSearch";
-import {
-  type AgentGraphDataResponse,
-  AgentGraphDataSchema,
-} from "@/src/features/trace-graph-view/types";
+import { type AgentGraphDataResponse } from "@/src/features/trace-graph-view/types";
+import { mapAgentGraphRecord } from "@/src/features/trace-graph-view/server/mapAgentGraphRecord";
 import { env } from "@/src/env.mjs";
 import {
   toDomainWithStringifiedMetadata,
@@ -703,42 +701,7 @@ export const traceRouter = createTRPCRouter({
       });
 
       const result = records
-        .map((r) => {
-          const parsed = AgentGraphDataSchema.safeParse(r);
-          if (!parsed.success) {
-            return null;
-          }
-
-          const data = parsed.data;
-          const hasLangGraphData = data.step != null && data.node != null;
-          const hasAgentData = data.type !== "EVENT"; // Include all types except EVENT
-
-          if (hasLangGraphData) {
-            return {
-              id: data.id,
-              node: data.node,
-              step: data.step,
-              parentObservationId: data.parent_observation_id || null,
-              name: data.name,
-              startTime: data.start_time,
-              endTime: data.end_time || undefined,
-              observationType: data.type,
-            };
-          } else if (hasAgentData) {
-            return {
-              id: data.id,
-              node: data.name,
-              step: 0,
-              parentObservationId: data.parent_observation_id || null,
-              name: data.name,
-              startTime: data.start_time,
-              endTime: data.end_time || undefined,
-              observationType: data.type,
-            };
-          }
-
-          return null;
-        })
+        .map((record) => mapAgentGraphRecord(record, "trace"))
         .filter((r) => Boolean(r)) as Required<AgentGraphDataResponse>[];
 
       return result;
